+++ src/uvm/refimpl/mem/MemConstants.java
@@ -1,7 +1,7 @@
 package uvm.refimpl.mem;
 
 public class MemConstants {
-    public static final int WORD_SIZE_LOG = 6;
-    public static final int WORD_SIZE_BITS = 1<<WORD_SIZE_LOG;
-    public static final int WORD_SIZE_BYTES = 1<<(WORD_SIZE_LOG-3);
+    public static final long WORD_SIZE_LOG = 6L;
+    public static final long WORD_SIZE_BITS = 1L << WORD_SIZE_LOG;
+    public static final long WORD_SIZE_BYTES = 1L << (WORD_SIZE_LOG - 3L);
 }
